Update operatii20210722-141629.cpp
Mathematical exercises. Enter the type of operation, the number of exercises and till what number  should be the operatin. Enter the result and it will be displayed if you answered correctly. If you made a mistake, the correct answers are displayed. Create a text file and write the results and date in it.